CI: fix failing server builds and also build the client for Linux and Android (#305)
* CI: Fix Basis Server Builds

Minor changes to make dotnet publish work in CI.
Remove extra jobs.

* CI: Refactor in prep for linux and android builds

* CI: Increase available disk space

* CI: Automatic Versioning

* CI: Build for Linux

This does dirty the tree, but only after we've extracted version
information. Not ideal, but I don't know of a better way to do it.

* CI: Build for Android

* CI: only restore cache from correct platform

given that cache for each platform takes at *minimum* 2GB,
I suspect that this will help by not accidentally duplicating
a platform's cache.

* CI: Only save cache on developer branch

* CI: don't use globs during artifact upload

They aren't having the desired effect, and I'm not sure why,
but this works correctly.

* CI: Redo release flow

* Set androidVersionCode in editor to match CI

* Document CI Secrets